Assignment 03/requirements.txt
confluent-kafka==2.5.3
pyarrow==16.1.0
pandas==2.2.2
